+++ src/components/datagrid/datagrid.js
@@ -7326,6 +7326,10 @@ Datagrid.prototype = {
         menu.append(`<li><a href="#" data-option="export-to-excel">${Locale.translate('ExportToExcel')}</a></li>`);
       }
 
+      if (this.settings.toolbar.exportToCsv) {
+        menu.append(`<li><a href="#" data-option="export-to-csv">${Locale.translate('ExportToCsv')}</a></li>`);
+      }
+
       if (this.settings.toolbar.advancedFilter) {
         menu.append(`<li><a href="#">${Locale.translate('AdvancedFilter')}</a></li>`);
       }
@@ -7414,6 +7418,10 @@ Datagrid.prototype = {
 
       if (action === 'export-to-excel') {
         self.exportToExcel();
+      }
+
+      if (action === 'export-to-csv') {
+        self.exportToCsv();
       }
 
       // Filter actions